Files: Moved server entry point to rootDir
